+++ examples/tina-cloud-starter/pages/posts/[filename].tsx
@@ -36,7 +36,7 @@ export const getStaticProps = async ({ params }) => {
         }
       }
     `,
-    variables: { relativePath: `${params.filename}.md` },
+    variables: { relativePath: `${params.filename}.mdx` },
   })) as { data: { getPostsDocument: PostsDocument } };
   return {
     props: {