+++ packages/@tinacms/toolkit/src/packages/fields/components/Toggle.tsx
@@ -96,7 +96,6 @@ const ToggleLabel = styled.label<{
   disabled?: boolean
 }>`
   background: none;
-  color: inherit;
   padding: 0;
   opacity: ${(props) => (props.disabled ? '0.4' : '1')};
   outline: none;